# Fix handling of surrogates on encoding