@JustAnotherArchivist 2022-04-17
Regarding the change of the `JSON_EncodeObject` interface, I figured that this is the cleanest way to implement it. However, if it's better avoided (is that considered public API? I'm not really sure...), the information is also available in the encoder struct, so it could be calculated from that in `objToJSON` instead.